requirements.txt
Django>=4.2,<5.0
mysqlclient>=2.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
python-dotenv>=1.0.0
djangorestframework-simplejwt==5.5.0
django-ckeditor==6.7.3